refactor(api): split check-in-record.service.ts into services/check-in-record/ (api-av1h)
