+++ .beans/api-av1h--refactor-check-in-recordservicets-590-loc-into-ser.md
@@ -1,10 +1,11 @@
 ---
 # api-av1h
 title: Refactor check-in-record.service.ts (590 LOC) into services/check-in-record/
-status: todo
+status: in-progress
 type: task
+priority: normal
 created_at: 2026-04-21T13:56:56Z
-updated_at: 2026-04-21T13:56:56Z
+updated_at: 2026-04-21T21:23:44Z
 parent: api-6l1q
 ---
 
@@ -39,3 +40,10 @@ Currently concentrates timer-triggered check-ins / response lifecycle / dismissa
 ## Parallelization
 
 No cross-blockers with other service refactor beans — safe to run in a worktree agent concurrently with siblings.
+
+## Findings
+
+- apps/api/src/services/check-in-record.service.ts — split 590 LOC monolith into 9 verb files under services/check-in-record/ (no barrel) — INFO
+- shared helpers toCheckInRecordResult (used across create/list/get/respond/dismiss/restore) and fetchPendingCheckIn (used by respond+dismiss) placed in internal.ts — INFO
+- CheckInRecordListOptions, CheckInRecordResult types moved to internal.ts and re-exported from list.ts as types — INFO
+- 8 callers updated (3 tests mocking the service needed vi.mock path fan-out to per-verb files) — INFO
