+++ apps/api/src/__tests__/routes/check-in-records/crud.test.ts
@@ -12,15 +12,27 @@ import type { ApiErrorResponse } from "@pluralscape/types";
 
 // ── Mocks ────────────────────────────────────────────────────────
 
-vi.mock("../../../services/check-in-record.service.js", () => ({
+vi.mock("../../../services/check-in-record/create.js", () => ({
   createCheckInRecord: vi.fn(),
+}));
+vi.mock("../../../services/check-in-record/list.js", () => ({
   listCheckInRecords: vi.fn(),
-  getCheckInRecord: vi.fn(),
-  respondCheckInRecord: vi.fn(),
-  dismissCheckInRecord: vi.fn(),
-  archiveCheckInRecord: vi.fn(),
-  deleteCheckInRecord: vi.fn(),
   parseCheckInRecordQuery: vi.fn().mockReturnValue({}),
+}));
+vi.mock("../../../services/check-in-record/get.js", () => ({
+  getCheckInRecord: vi.fn(),
+}));
+vi.mock("../../../services/check-in-record/respond.js", () => ({
+  respondCheckInRecord: vi.fn(),
+}));
+vi.mock("../../../services/check-in-record/dismiss.js", () => ({
+  dismissCheckInRecord: vi.fn(),
+}));
+vi.mock("../../../services/check-in-record/archive.js", () => ({
+  archiveCheckInRecord: vi.fn(),
+}));
+vi.mock("../../../services/check-in-record/delete.js", () => ({
+  deleteCheckInRecord: vi.fn(),
 }));
 
 vi.mock("../../../lib/audit-writer.js", () => mockAuditWriterFactory());
@@ -32,15 +44,13 @@ vi.mock("../../../middleware/rate-limit.js", () => mockRateLimitFactory());
 vi.mock("../../../middleware/auth.js", () => mockAuthFactory());
 // ── Imports after mocks ──────────────────────────────────────────
 
-const {
-  createCheckInRecord,
-  listCheckInRecords,
-  getCheckInRecord,
-  respondCheckInRecord,
-  dismissCheckInRecord,
-  archiveCheckInRecord,
-  deleteCheckInRecord,
-} = await import("../../../services/check-in-record.service.js");
+const { createCheckInRecord } = await import("../../../services/check-in-record/create.js");
+const { listCheckInRecords } = await import("../../../services/check-in-record/list.js");
+const { getCheckInRecord } = await import("../../../services/check-in-record/get.js");
+const { respondCheckInRecord } = await import("../../../services/check-in-record/respond.js");
+const { dismissCheckInRecord } = await import("../../../services/check-in-record/dismiss.js");
+const { archiveCheckInRecord } = await import("../../../services/check-in-record/archive.js");
+const { deleteCheckInRecord } = await import("../../../services/check-in-record/delete.js");
 const { createCategoryRateLimiter } = await import("../../../middleware/rate-limit.js");
 const { systemRoutes } = await import("../../../routes/systems/index.js");
