+++ apps/api/src/routes/check-in-records/delete.ts
@@ -6,7 +6,7 @@ import { createAuditWriter } from "../../lib/audit-writer.js";
 import { getDb } from "../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { deleteCheckInRecord } from "../../services/check-in-record.service.js";
+import { deleteCheckInRecord } from "../../services/check-in-record/delete.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
