+++ apps/api/src/routes/check-in-records/dismiss.ts
@@ -6,7 +6,7 @@ import { getDb } from "../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { dismissCheckInRecord } from "../../services/check-in-record.service.js";
+import { dismissCheckInRecord } from "../../services/check-in-record/dismiss.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
