+++ apps/api/src/routes/check-in-records/get.ts
@@ -5,7 +5,7 @@ import { getDb } from "../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { getCheckInRecord } from "../../services/check-in-record.service.js";
+import { getCheckInRecord } from "../../services/check-in-record/get.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
