+++ apps/api/src/routes/check-in-records/respond.ts
@@ -7,7 +7,7 @@ import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { parseJsonBody } from "../../lib/parse-json-body.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { respondCheckInRecord } from "../../services/check-in-record.service.js";
+import { respondCheckInRecord } from "../../services/check-in-record/respond.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
