+++ apps/api/src/routes/check-in-records/restore.ts
@@ -6,7 +6,7 @@ import { getDb } from "../../lib/db.js";
 import { parseIdParam, requireIdParam } from "../../lib/id-param.js";
 import { envelope } from "../../lib/response.js";
 import { createCategoryRateLimiter } from "../../middleware/rate-limit.js";
-import { restoreCheckInRecord } from "../../services/check-in-record.service.js";
+import { restoreCheckInRecord } from "../../services/check-in-record/restore.js";
 
 import type { AuthEnv } from "../../lib/auth-context.js";
