+++ apps/api/src/services/check-in-record/delete.ts
@@ -1,0 +1,46 @@
+import { checkInRecords } from "@pluralscape/db/pg";
+import { and, eq } from "drizzle-orm";
+
+import { HTTP_NOT_FOUND } from "../../http.constants.js";
+import { ApiHttpError } from "../../lib/api-error.js";
+import { withTenantTransaction } from "../../lib/rls-context.js";
+import { assertSystemOwnership } from "../../lib/system-ownership.js";
+import { tenantCtx } from "../../lib/tenant-context.js";
+
+import type { AuditWriter } from "../../lib/audit-writer.js";
+import type { AuthContext } from "../../lib/auth-context.js";
+import type { CheckInRecordId, SystemId } from "@pluralscape/types";
+import type { PostgresJsDatabase } from "drizzle-orm/postgres-js";
+
+export async function deleteCheckInRecord(
+  db: PostgresJsDatabase,
+  systemId: SystemId,
+  recordId: CheckInRecordId,
+  auth: AuthContext,
+  audit: AuditWriter,
+): Promise<void> {
+  assertSystemOwnership(systemId, auth);
+
+  await withTenantTransaction(db, tenantCtx(systemId, auth), async (tx) => {
+    const [existing] = await tx
+      .select({ id: checkInRecords.id })
+      .from(checkInRecords)
+      .where(and(eq(checkInRecords.id, recordId), eq(checkInRecords.systemId, systemId)))
+      .limit(1);
+
+    if (!existing) {
+      throw new ApiHttpError(HTTP_NOT_FOUND, "NOT_FOUND", "Check-in record not found");
+    }
+
+    await tx
+      .delete(checkInRecords)
+      .where(and(eq(checkInRecords.id, recordId), eq(checkInRecords.systemId, systemId)));
+
+    await audit(tx, {
+      eventType: "check-in-record.deleted",
+      actor: { kind: "account", id: auth.accountId },
+      detail: "Check-in record deleted",
+      systemId,
+    });
+  });
+}
